fix 修复角色、岗位、字典数据在PostgreSQL执行插入或者更新SQL报错问题，修改对应实体类的排序属性为Integer类型。

+++ ruoyi-common/src/main/java/com/ruoyi/common/core/domain/entity/SysDictData.java
@@ -42,7 +42,7 @@ public class SysDictData extends BaseEntity {
      */
     @ApiModelProperty(value = "字典排序")
     @ExcelProperty(value = "字典排序")
-    private Long dictSort;
+    private Integer dictSort;
 
     /**
      * 字典标签
